implémentation du chargement/enregistrement d'une scène

+++ egs3d/trunk/org.egs3d.core.resources/src/org/egs3d/core/resources/internal/SceneReader.java
@@ -27,17 +27,27 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
 import javax.media.j3d.BranchGroup;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathFactory;
 
 import org.egs3d.core.resources.IBranchGroupContainer;
+import org.egs3d.core.resources.IModel;
 import org.egs3d.core.resources.IModelContainer;
 import org.egs3d.core.resources.IScene;
 import org.egs3d.core.resources.ISceneReader;
+import org.egs3d.core.resources.ITexture;
 import org.egs3d.core.resources.ITextureContainer;
 import org.egs3d.core.resources.ResourcesPlugin;
+import org.w3c.dom.Element;
+import org.w3c.dom.NodeList;
+import org.xml.sax.InputSource;
 
+import com.sun.j3d.loaders.Loader;
 import com.sun.j3d.utils.scenegraph.io.SceneGraphFileReader;
 
 
@@ -78,8 +88,11 @@ public class SceneReader implements ISceneReader {
 
     private void readBranchGroups(ZipFile zipFile, IBranchGroupContainer container)
             throws Exception {
-        final InputStream input = zipFile.getInputStream(zipFile
-                .getEntry(SceneIOConstants.SCENE_NAME));
+        final ZipEntry entry = zipFile.getEntry(SceneIOConstants.SCENE_NAME);
+        if (entry == null) {
+            return;
+        }
+        final InputStream input = zipFile.getInputStream(entry);
 
         final File tempFile = File.createTempFile("scene-", ".tmp");
         tempFile.deleteOnExit();
@@ -103,13 +116,80 @@ public class SceneReader implements ISceneReader {
     }
 
 
+    @SuppressWarnings("unchecked")
     private void readModels(ZipFile zipFile, IModelContainer container) throws Exception {
-        // TODO implémentation
+        final ZipEntry descEntry = zipFile.getEntry(SceneIOConstants.DESCRIPTOR_NAME);
+        if (descEntry == null) {
+            // pas de descripteur
+            return;
+        }
+
+        final InputStream input = zipFile.getInputStream(descEntry);
+        final XPath xpath = XPathFactory.newInstance().newXPath();
+        final NodeList nodeList = (NodeList) xpath.evaluate("/egs/models/model",
+                new InputSource(input), XPathConstants.NODESET);
+        final int nodeListLen = nodeList.getLength();
+
+        for (int i = 0; i < nodeListLen; ++i) {
+            final Element elem = (Element) nodeList.item(i);
+            final String name = elem.getAttribute("name");
+            final String file = elem.getAttribute("file");
+            final String loaderClassName = elem.getAttribute("loaderClass");
+
+            if (name.length() == 0 || file.length() == 0 || loaderClassName.length() == 0) {
+                // pas d'attribut "name" ou "file" ou "loaderClass"
+                continue;
+            }
+
+            final ZipEntry modelEntry = zipFile.getEntry(SceneIOConstants.MODELS_NAME
+                    + file);
+            if (modelEntry != null) {
+                final Class<? extends Loader> loaderClass;
+                try {
+                    loaderClass = (Class<? extends Loader>) Class
+                            .forName(loaderClassName);
+                } catch (ClassNotFoundException e) {
+                    continue;
+                }
+                final IModel model = container.create(name, loaderClass, zipFile
+                        .getInputStream(modelEntry));
+                container.add(model);
+            }
+        }
     }
 
 
     private void readTextures(ZipFile zipFile, ITextureContainer container)
             throws Exception {
-        // TODO implémentation
+        final ZipEntry descEntry = zipFile.getEntry(SceneIOConstants.DESCRIPTOR_NAME);
+        if (descEntry == null) {
+            // pas de descripteur
+            return;
+        }
+
+        final InputStream input = zipFile.getInputStream(descEntry);
+        final XPath xpath = XPathFactory.newInstance().newXPath();
+        final NodeList nodeList = (NodeList) xpath.evaluate("/egs/textures/texture",
+                new InputSource(input), XPathConstants.NODESET);
+        final int nodeListLen = nodeList.getLength();
+
+        for (int i = 0; i < nodeListLen; ++i) {
+            final Element elem = (Element) nodeList.item(i);
+            final String name = elem.getAttribute("name");
+            final String file = elem.getAttribute("file");
+
+            if (name.length() == 0 || file.length() == 0) {
+                // pas d'attribut "name" ou "file"
+                continue;
+            }
+
+            final ZipEntry textureEntry = zipFile.getEntry(SceneIOConstants.TEXTURES_NAME
+                    + file);
+            if (textureEntry != null) {
+                final ITexture texture = container.create(name, zipFile
+                        .getInputStream(descEntry));
+                container.add(texture);
+            }
+        }
     }
 }
